mvc 구조 적용
