+++ server/mongoose/schema/board.js
@@ -1,9 +1,11 @@
 const mongoose = require('mongoose');
 const Schema = mongoose.Schema;
 
-const Board = new Schema({
+const BoardSchema = new Schema({
   title: { type: String, required: true },
   slug: { type: String, required: true, unique: true },
 }, { timestamps: true });
+
+const Board = mongoose.model("Board", BoardSchema)
 
 module.exports = Board;
