+++ server/server.js
@@ -1,6 +1,8 @@
 const express = require("express");
 const cors = require("cors");
 const app = express();
+const mongoose = require('mongoose');
+require("dotenv").config();
 
 const { article, board, comment, reply, user } = require("./router");
 const PORT = 8080;
@@ -20,6 +22,26 @@ app.use(board);
 app.use(comment);
 app.use(reply);
 app.use(user);
+
+const db = mongoose.connection;
+
+(() => {
+  try {
+    // db.on
+    db.on('error', console.error);
+    db.on('open', () => {
+      console.log(`connected mongoDB!`);
+    });
+
+    // mongoDB connect
+    mongoose.connect(
+      `mongodb+srv://${process.env.DB_ID}:${process.env.DB_PASSWORD}@cluster0.gnquy.mongodb.net/refactoringProject?retryWrites=true&w=majority`
+    );
+
+  } catch (error) {
+    console.log({ error: error.message });
+  }
+})();
 
 const html = `
 <!DOCTYPE html>
@@ -42,7 +64,9 @@ app.get("/", (req, res) => {
 });
 
 if (process.env.NODE_ENV == 'development') {
-  console.log('development!');
+  app.listen(PORT, () => {
+    console.log(`development App listen is ${PORT}`);
+  });
 } else if (process.env.NODE_ENV == 'test') {
   console.log('test!');
 } else {
